리프레쉬 토큰 관련 보완

+++ src/main/java/com/my/bob/member/entity/BobUserRefreshToken.java
@@ -16,7 +16,7 @@ public class BobUserRefreshToken {
     @Id
     @Column(name = "TOKEN_ID")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int tokenId;
+    private long tokenId;
 
     private long userId;
